Replace merge_paths with Path::Tiny:
This was a bit hard to figure out, but with the help of @wchristian,
it seems as though this merges path from argument 1 onto argument 2.

I have this replaced it with a single path() call. All tests pass,
so I hope this is okay. We should run these tests under Windows to
make sure it works there as well.